move helper classes into specific folder
